Refactor contribution form to use shared Form component
- Create a new Form component to centralize contribution form logic
- Extract common form handling, validation, and UI elements from New and Edit pages
- Simplify form rendering by using a reusable component with flexible props
- Improve code organization and reduce duplication in contribution form handling
